Fixes for gcc

+++ string_view.hpp
@@ -15,7 +15,7 @@ FORCE_INLINE
 bool32 operator == (string_view s1, string_view s2)
 {
     bool32 is_equal = (s1.size == s2.size);
-    for (int i = 0; is_equal && i < s1.size; i++)
+    for (usize i = 0; is_equal && i < s1.size; i++)
     {
         is_equal = (s1.data[i] == s2.data[i]);
     }
